move native test to example
